test(correlation): revive settings-page + alert-grouping coverage against new UI
The old GeneralTests/correlationSettings.spec.js (13 tests) targeted the
pre-revamp Settings UI and was disabled. Revive the still-valid coverage,
consolidated into the Correlation suite:

- Page object fixes: tab labels renamed (Configuration->Detection Rules,
  Alert Correlation->Alert Grouping, Services->Discovered Services);
  expectServiceDiscoveryContentVisible anchors on the new 'Save Configuration'
  button (old dual-list save-btn removed); ensureSemanticGroupsExist drops the
  removed per-group flags (is_scope/is_stable/normalize); add clickDiscoveredServicesTab.
- correlation-settings.spec.js: + navigation (4 tabs, switching) + Discovered
  Services tab (refresh/loading) — shared-org, navigateToBase.
- correlation-alert-grouping.spec.js (new): 7 deduplication tests (distinct
  feature — a coverage gap the running suite never touched).
- Drop the 2 stale Configuration-tab tests (old dual-list picker is gone;
  superseded by SETTINGS-01 / CACHE-02).
- Delete GeneralTests/correlationSettings.spec.js.

12 tests collect under playwright.config.js. Validated by the ENT run.
