+++ tests/ui-testing/pages/generalPages/correlationSettingsPage.js
@@ -28,9 +28,12 @@ export class CorrelationSettingsPage {
         this.saveConfigurationButtonName = 'Save Configuration';
         // Placeholder text of the "+ Add field" OSelect trigger.
         this.selectAFieldText = 'Select a field';
-        this.servicesTabName = 'Services';
-        this.serviceDiscoveryTabName = 'Configuration';
-        this.alertCorrelationTabName = 'Alert Correlation';
+        // Visible tab labels (post-revamp): the tab `name` attrs are
+        // services / discovery / alert-correlation / field-aliases, rendered as
+        // these i18n labels.
+        this.servicesTabName = 'Discovered Services';
+        this.serviceDiscoveryTabName = 'Detection Rules';
+        this.alertCorrelationTabName = 'Alert Grouping';
         this.fieldAliasesTabName = 'Field Mappings';
         // Detection Rules view (the "Distinguish Services By" configuration tab)
         this.detectionRulesTabName = 'Detection Rules';
@@ -92,43 +95,33 @@ export class CorrelationSettingsPage {
         const password = process.env["ZO_ROOT_USER_PASSWORD"];
         const authHeader = 'Basic ' + Buffer.from(`${username}:${password}`).toString('base64');
 
-        // Define test semantic groups that cover common use cases
+        // Define test semantic groups that cover common use cases.
+        // Post-revamp shape: {id, display, group, fields} only — the per-group
+        // flags (is_scope / is_stable / normalize) were removed.
         const testSemanticGroups = [
             {
                 id: 'k8s-cluster',
                 display: 'K8s Cluster',
                 group: 'Kubernetes',
                 fields: ['k8s.cluster.name', 'k8s_cluster_name'],
-                normalize: true,
-                is_scope: true,
-                is_stable: true
             },
             {
                 id: 'k8s-namespace',
                 display: 'K8s Namespace',
                 group: 'Kubernetes',
                 fields: ['k8s.namespace.name', 'k8s_namespace_name'],
-                normalize: true,
-                is_scope: true,
-                is_stable: true
             },
             {
                 id: 'k8s-deployment',
                 display: 'K8s Deployment',
                 group: 'Kubernetes',
                 fields: ['k8s.deployment.name', 'k8s_deployment_name'],
-                normalize: true,
-                is_scope: false,
-                is_stable: true
             },
             {
                 id: 'service',
                 display: 'Service',
                 group: 'Common',
                 fields: ['service.name', 'service_name', 'service'],
-                normalize: true,
-                is_scope: false,
-                is_stable: true
             }
         ];
 
@@ -209,6 +202,11 @@ export class CorrelationSettingsPage {
         const tab = this.page.getByRole('tab', { name: this.servicesTabName });
         await tab.click();
         await this.page.waitForLoadState('networkidle', { timeout: 10000 }).catch(() => {});
+    }
+
+    // The "services" tab IS the Discovered Services tab (post-revamp label).
+    async clickDiscoveredServicesTab() {
+        await this.clickServicesTab();
     }
 
     async clickAlertCorrelationTab() {
@@ -364,8 +362,11 @@ export class CorrelationSettingsPage {
     }
 
     async expectServiceDiscoveryContentVisible() {
-        // Wait for the Configuration tab content to be visible (formerly Service Identity)
-        await expect(this.page.locator(this.serviceIdentitySaveBtn)).toBeVisible({ timeout: 15000 });
+        // Detection Rules view: the "Save Configuration" button anchors it
+        // (the old dual-list Service Identity save-btn was removed in the revamp).
+        await expect(
+            this.page.getByRole('button', { name: this.saveConfigurationButtonName }).first(),
+        ).toBeVisible({ timeout: 15000 });
     }
 
     async expectFieldAliasesContentVisible() {
